# Optimize HTMLTree.build