@philss 2023-12-24
@ypconstante thank you very much for the improvements you are proposing! :purple_heart: 
I should review all of them this week!